hapus lahan_image controller and service
