+++ controller/__init__.py
@@ -1,6 +1,5 @@
 __all__ = [
     "lahan_controller",
-    "lahan_image_controller",
     "user_controller",
     "auth_controller",
 ]
